Rename api and file service files
